pyproject.toml
[tool.poetry]
name = "blog-project"
version = "0.1.0"
description = ""
authors = ["Andrei <sharik.and@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "3"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"
drf-yasg = "^1.20.0"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
